Extract Google Sign In logic
