+++ src/legacy/screen.js
@@ -12,6 +12,7 @@ import {
   FormatTicketDateTime,
   FormatPrice,
 } from "./modules/products";
+import { setupLoginHandling, showLoggedOutState } from "./modules/googleAuth";
 // Inject channelId into window for any non-React scripts
 (function() {
   const DEFAULT_CHANNEL_ID =
@@ -1133,229 +1134,6 @@ function populateFavoritesTab() {
   console.log(`Added ${upvotedItems.length} items to the favorites tab.`);
 }
 
-// This bit handles logging in
-function setupLoginHandling() {
-  // --- Configuration ---
-  const backendBaseUrl = "https://fastapi.edgevideo.ai";
-  const authRouteBase = `${backendBaseUrl}/auth_google`;
-  const userInfoUrl = `${authRouteBase}/details`; // URL to get user details
-  const frontendUrl = window.location.origin + window.location.pathname; // URL for redirect
-
-  // --- UI Element Selection (New Page Structure) ---
-  // Use querySelector as ID wasn't provided and it might be the only one
-  const signInButton = document.querySelector("button.gsi-material-button");
-  const logoutDiv = document.getElementById("logout"); // The logout 'button' is now a div
-  const usernameElements = document.querySelectorAll(".username"); // All elements to show the username
-  const avatarContainers = document.querySelectorAll(".avatar-container");
-  const avatarImages = document.querySelectorAll(".avatar"); // All avatar image elements
-  const errorMessageP = document.getElementById("errorMessage"); // Optional: For displaying errors
-  const favoritesButton = document.getElementById("favoritesBtn");
-
-  // --- Function to Fetch User Details (Mostly unchanged from original) ---
-  async function fetchUserDetails(token) {
-    if (errorMessageP) errorMessageP.textContent = ""; // Clear previous errors
-    try {
-      const response = await fetch(userInfoUrl, {
-        method: "GET",
-        headers: {
-          Authorization: `Bearer ${token}`, // Send the JWT
-        },
-      });
-
-      if (response.status === 401) {
-        console.error("Token is invalid or expired.");
-        localStorage.removeItem("authToken");
-        let errorJson = null;
-        try {
-          errorJson = await response.json();
-        } catch (e) {
-          /* ignore */
-        }
-        const message =
-          errorJson?.message || "Session invalid. Please log in again.";
-        throw new Error(message);
-      }
-      if (!response.ok) {
-        let errorJson = null;
-        try {
-          errorJson = await response.json();
-        } catch (e) {
-          /* ignore */
-        }
-        const message =
-          errorJson?.message ||
-          `Failed to fetch user details (Status: ${response.status})`;
-        throw new Error(message);
-      }
-
-      const userData = await response.json();
-      // Use name, fallback to email, fallback to 'User'
-      const displayName = userData.name || userData.email || "User";
-      // Return all data plus the derived displayName
-      return { ...userData, displayName };
-    } catch (error) {
-      console.error("Error fetching user details:", error);
-      if (errorMessageP) {
-        errorMessageP.textContent =
-          error.message || "Could not retrieve user information.";
-      }
-      showLoggedOutState(); // Log out UI on fetch failure
-      return null;
-    }
-  }
-
-  // --- UI Update Functions (Adapted for New Page Structure) ---
-  async function showLoggedInState(currentToken) {
-    const user = await fetchUserDetails(currentToken);
-
-    if (user) {
-      // 1. Hide Sign-In Button
-      if (signInButton) {
-        signInButton.style.display = "none";
-      }
-
-      // 2. Show Log Out Div
-      if (logoutDiv) {
-        logoutDiv.style.display = "flex"; // Or 'flex', 'inline-block' etc. as needed
-      }
-
-      // 3. Update Username Elements
-      usernameElements.forEach((el) => {
-        el.textContent = user.displayName;
-      });
-
-      // 4. Update Avatar Images
-      const avatarSeed = encodeURIComponent(user.displayName); // Ensure name is URL-safe
-      const avatarUrl = `https://api.dicebear.com/9.x/bottts/svg?seed=${avatarSeed}`;
-      avatarImages.forEach((img) => {
-        img.src = avatarUrl;
-        img.alt = `${user.displayName}'s avatar`; // Good practice for accessibility
-        img.style.display = "flex";
-      });
-      avatarContainers.forEach((div) => {
-        div.style.display = "flex";
-      });
-
-      // Optional: Clear any residual error messages
-      if (errorMessageP) errorMessageP.textContent = "";
-
-      console.log(`Logged in as: ${user.displayName} (${user.email})`);
-
-      await fetchVotedProducts();
-    }
-    // If fetchUserDetails failed, it already called showLoggedOutState()
-  }
-
-  function showLoggedOutState() {
-    // 1. Show Sign-In Button
-    if (signInButton) {
-      // Make sure it's visible. 'block' is a common default. Adjust if needed.
-      signInButton.style.display = "flex";
-    }
-
-    // 2. Hide Log Out Div
-    if (logoutDiv) {
-      logoutDiv.style.display = "none";
-    }
-
-    // 3. Clear Username Elements
-    usernameElements.forEach((el) => {
-      el.textContent = ""; // Or set to a default like 'Guest'
-    });
-
-    // 4. Reset or set default Avatar Images
-    const defaultAvatarUrl =
-      "https://api.dicebear.com/9.x/bottts/svg?seed=guest"; // Default avatar
-    avatarImages.forEach((img) => {
-      img.src = defaultAvatarUrl;
-      img.alt = "Default avatar";
-    });
-    avatarContainers.forEach((div) => {
-      div.style.display = "none";
-    });
-
-    // Optional: Display logged-out message (don't clear potential important errors)
-    // if (errorMessageP && !errorMessageP.textContent) { // Only set if no specific error shown
-    //    errorMessageP.textContent = 'You are logged out.';
-    // }
-    console.log("Showing logged out state.");
-
-    votedProducts = [];
-  }
-
-  // --- Event Listeners (Attached to New Elements) ---
-  if (signInButton) {
-    signInButton.addEventListener("click", () => {
-      if (errorMessageP) errorMessageP.textContent = ""; // Clear errors
-      // Redirect to Google login endpoint
-      const googleLoginUrl = `${authRouteBase}/google?redirectUri=${encodeURIComponent(
-        frontendUrl
-      )}`;
-      window.location.href = googleLoginUrl;
-    });
-  } else {
-    console.warn('Sign in button with class "gsi-material-button" not found.');
-  }
-
-  if (logoutDiv) {
-    logoutDiv.addEventListener("click", () => {
-      if (errorMessageP) errorMessageP.textContent = ""; // Clear errors
-      localStorage.removeItem("authToken");
-      showLoggedOutState();
-    });
-    // Add cursor pointer to make the div behave more like a button visually
-    logoutDiv.style.cursor = "pointer";
-  } else {
-    console.warn('Logout div with ID "logout" not found.');
-  }
-
-  if (favoritesButton) {
-    favoritesButton.addEventListener("click", () => {
-      console.log("Favorites button clicked.");
-      // Check if logged in before populating? Optional, depends on UX.
-      const token = localStorage.getItem("authToken");
-      if (token) {
-        populateFavoritesTab();
-      } else {
-        console.log("User not logged in, cannot show favorites.");
-        // Optional: Show a message asking the user to log in.
-        const favsContainer = document.getElementById("favs");
-        if (favsContainer) {
-          favsContainer
-            .querySelectorAll(".fav-item:not(.none)")
-            .forEach((item) => item.remove());
-          // Display login message if desired
-        }
-      }
-    });
-  } else {
-    console.warn("Favorites button with ID 'favoritesBtn' not found.");
-  }
-
-  // --- Initial Page Load Logic (Unchanged from original) ---
-  const urlParams = new URLSearchParams(window.location.search);
-  const tokenFromUrl = urlParams.get("token");
-  let currentToken = null;
-
-  if (tokenFromUrl) {
-    currentToken = tokenFromUrl;
-    localStorage.setItem("authToken", currentToken);
-    // Clean the URL
-    window.history.replaceState({}, document.title, window.location.pathname);
-    console.log("Token received from URL, attempting login...");
-    showLoggedInState(currentToken); // Update UI based on token
-  } else {
-    const storedToken = localStorage.getItem("authToken");
-    if (storedToken) {
-      console.log("Token found in localStorage, attempting login...");
-      currentToken = storedToken;
-      showLoggedInState(currentToken); // Update UI based on stored token
-    } else {
-      console.log("No token found. Showing logged out state.");
-      showLoggedOutState(); // Show initial logged-out state
-    }
-  }
-}
 
 // --- NEW: Main Initialization Function ---
 function initializeApp() {
@@ -1372,7 +1150,7 @@ function initializeApp() {
   setInterval(UpdateFaces, 1000);
 
   // Set up the login/auth related UI and logic
-  setupLoginHandling();
+  setupLoginHandling(fetchVotedProducts, populateFavoritesTab);
 
   // Any other initialization code that depends on the DOM should go here
   hideConsoleMessages(); // Call this after DOM ready if it depends on anything
